refactor: isolate the discord components
