+++ quote_bot/discord_components/client.py
@@ -1,0 +1,11 @@
+import discord
+from discord.ext import commands
+
+intents = discord.Intents.default()
+intents.message_content = True
+client = commands.Bot(command_prefix="/", intents=intents)
+
+
+@client.event
+async def on_ready() -> None:
+    print(f"We have logged in as {client.user}")
